// cabineverde/src/types/enums.ts
export enum FaixaEtaria {
  CRIANCA = 'Criança',
  PRE_ADOLESCENTE = 'Pré-adolescente',
  ADOLESCENTE = 'Adolescente',
  ADULTO = 'Adulto',
  IDOSO = 'Idoso'
}

export enum ClassificacaoRisco {
  ALTO = 'Alto risco',
  MODERADO = 'Risco moderado',
  BAIXO = 'Baixo risco'
}

export enum Prioridade {
  CRITICA = 'Crítica',
  ALTA = 'Alta',
  MEDIA = 'Média',
  BAIXA = 'Baixa'
}

export enum StatusCaso {
  EM_TRIAGEM = 'Em triagem',
  EM_BUSCA = 'Em busca',
  LOCALIZADO = 'Localizado',
  ENCERRADO = 'Encerrado'
}